feature: created firefighter-reports type for frontend refactor
