+++ app/frontend/src/types/firefighter-reports.ts
@@ -1,0 +1,31 @@
+import type { ReportStatus } from './report';
+
+export interface FirefighterReportTable {
+    ref: string;
+    location: string;
+    status: ReportStatus;
+    size: number;
+    reported: string;
+    reporter: string;
+    lat: number;
+    lng: number;
+}
+
+export interface FirefighterReportModal {
+    id: string;
+    ref: string;
+    location: string;
+    status: ReportStatus;
+    reported: string;
+    reporter: string;
+    description: string;
+    image_url: string;
+    size: number;
+    lat: number;
+    lng: number;
+}
+
+export interface ReportList {
+    data: FirefighterReportTable[];
+    total: number;
+}
